+++ SubModule.xml
@@ -1,7 +1,7 @@
 <Module>
 	<Name value="Custom Troop Upgrade And Replacer Mod"/>
 	<Id value="CustomTroopUpgrades"/>
-	<Version value="v0.3.2"/>
+	<Version value="v0.4.0"/>
 	<SingleplayerModule value="true"/>
 	<MultiplayerModule value="false"/>
 	<Official value="false"/>
